vortex/pyproject.toml
[project]
name = "vortex"
version = "0.1.0"
description = "ICN Vortex Engine - GPU-resident AI video generation pipeline"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "ICN Team" }]

dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.36.0",
    "diffusers>=0.30.0",  # CogVideoX I2V support with PipelineQuantizationConfig
    "accelerate>=0.30.0", # CPU offload support for CogVideoX
    "safetensors>=0.4.0",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
    "opencv-python>=4.8.0",
    "soundfile>=0.12.0",
    "einops>=0.7.0",
    "bitsandbytes>=0.42.0",  # For NF4/INT8 quantization (CogVideoX)
    "torchao>=0.4.0",        # INT8 weight-only quantization for CogVideoX
    "nvidia-ml-py>=12.535.0", # VRAM monitoring (replaces deprecated pynvml)
    "prometheus-client>=0.19.0",
    "kokoro>=0.7.0",         # TTS model (T017) - KPipeline for TTS
    "pyyaml>=6.0.0",         # Config file loading (T017)
    "open-clip-torch>=2.23.0",  # CLIP ensemble semantic verification (T018)
    "av>=12.0.0",  # VP9/Opus encoding for Lane 0 video pipeline
    "librosa>=0.10.0",  # Audio analysis for audio-gated motion driver
    "scipy>=1.10.0",    # Signal smoothing for audio-gated motion driver
    "aiohttp>=3.9.0",   # Async HTTP client for ComfyUI API
    "websockets>=12.0", # WebSocket client for ComfyUI progress monitoring
    "httpx>=0.26.0",    # Async HTTP client for Ollama Showrunner API
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "matplotlib>=3.8.0",
    "scipy>=1.11.0",        # For audio processing tests (T017)
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vortex"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
